core code updates - updatable result set enhancements
git-svn-id: http://svn.code.sf.net/p/hsqldb/svn/base/trunk@6771 7c7dc5f5-a22d-0410-a3af-b41755a11667

+++ src/org/hsqldb/Row.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2001-2023, The HSQL Development Group
+/* Copyright (c) 2001-2024, The HSQL Development Group
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -43,7 +43,7 @@ import org.hsqldb.types.TimestampData;
  * Base class for a database row object.
  *
  * @author Fred Toussi (fredt@users dot sourceforge dot net)
- * @version 2.7.3
+ * @version 2.7.4
  */
 public class Row implements CachedObject {
 
@@ -168,7 +168,7 @@ public class Row implements CachedObject {
     }
 
     public long getId() {
-        return ((long) table.getId() << 40) + position;
+        return (table.getId() << 40) + position;
     }
 
     public void setPos(long pos) {
